ToolBox-Ver 1.27.2.5

+++ controllers/categorias.js
@@ -22,7 +22,6 @@ router.post('/', async (req, res) => {
         const nuevaCategoria = new Categoria({ nombre, numero: nuevoNumero });
         await nuevaCategoria.save();
 
-        // Redirige a la página de gestión de categorías con mensaje de éxito
         res.redirect('/inventario/categoria?success=Categoría agregada con éxito');
     } catch (error) {
         res.status(500).json({ error: 'Error al agregar categoría' });
@@ -30,33 +29,29 @@ router.post('/', async (req, res) => {
 });
 
 // Actualizar una categoría
-router.put('/:id', async (req, res) => {
+router.post('/:id', async (req, res) => {
     try {
         const { id } = req.params;
-        const { nombre } = req.body;
-        const categoriaActualizada = await Categoria.findByIdAndUpdate(id, { nombre }, { new: true });
-        if (!categoriaActualizada) {
-            return res.status(404).json({ error: 'Categoría no encontrada' });
-        }
-        // Redirige a la lista de categorías con mensaje de éxito
-        res.redirect('/inventario/categoria?success=Categoría actualizada con éxito');
-    } catch (error) {
-        res.status(500).json({ error: 'Error al actualizar la categoría' });
-    }
-});
+        const { _method } = req.body;
 
-// Eliminar una categoría
-router.delete('/:id', async (req, res) => {
-    try {
-        const { id } = req.params;
-        const categoriaEliminada = await Categoria.findByIdAndDelete(id);
-        if (!categoriaEliminada) {
-            return res.status(404).json({ error: 'Categoría no encontrada' });
+        if (_method === 'PUT') {
+            const { nombre } = req.body;
+            const categoriaActualizada = await Categoria.findByIdAndUpdate(id, { nombre }, { new: true });
+            if (!categoriaActualizada) {
+                return res.status(404).json({ error: 'Categoría no encontrada' });
+            }
+            res.redirect('/inventario/categoria?success=Categoría actualizada con éxito');
+        } else if (_method === 'DELETE') {
+            const categoriaEliminada = await Categoria.findByIdAndDelete(id);
+            if (!categoriaEliminada) {
+                return res.status(404).json({ error: 'Categoría no encontrada' });
+            }
+            res.redirect('/inventario/categoria?success=Categoría eliminada con éxito');
+        } else {
+            res.status(400).json({ error: 'Método no soportado' });
         }
-        // Redirige a la lista de categorías con mensaje de éxito
-        res.redirect('/inventario/categoria?success=Categoría eliminada con éxito');
     } catch (error) {
-        res.status(500).json({ error: 'Error al eliminar la categoría' });
+        res.status(500).json({ error: 'Error al procesar la solicitud' });
     }
 });
